fix: Phase 1 bug fixes — 6 correctness and safety improvements (BUG-24 through BUG-29)
- BUG-24: Fix YUM/DNF getPackageCaches() copy-paste bug — was calling
  getPacmanPackageCaches(). Added getYumDnfPackageCaches() scanning
  /var/cache/dnf/ and /var/cache/yum/
- BUG-25: Fix CircleBar double-delete — QChartView owns the QChart,
  removed manual delete from destructor
- BUG-26: Fix DiskInfo raw pointer ownership — changed QList<Disk*> to
  QList<Disk> value semantics, eliminating Rule of Three violation
- BUG-27: Fix /proc/meminfo bounds check — added guard for < 8 lines
- BUG-28: Fix quint8 core count overflow — changed to int (both platforms)
- BUG-29: Fix toLong() truncation — changed to toLongLong() for memory,
  network, and process values that can exceed 32-bit range

+++ linux/nexis-core/Info/cpu_info.cpp
@@ -43,7 +43,7 @@ int CpuInfo::getCpuPhysicalCoreCount() const
 
 int CpuInfo::getCpuCoreCount() const
 {
-    static quint8 count = 0;
+    static int count = 0;
 
     if (! count) {
         QStringList cpuinfo = FileUtil::readListFromFile(PROC_CPUINFO);
